requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cachetools==4.2.0
Cap1xxx==0.1.3
certifi==2020.12.5
chardet==4.0.0
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cupshelpers==1.0
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
fourletterphat==0.1.0
google-api-core==1.25.0
google-auth==1.24.0
google-cloud-pubsub==2.2.0
googleapis-common-protos==1.52.0
gpiozero==1.5.1
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
html5lib==1.0.1
idna==2.10
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
lazy-object-proxy==1.3.1
libcst==0.3.16
logilab-common==1.4.2
lxml==4.3.2
MarkupSafe==1.1.0
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.3
numpy==1.16.2
oauthlib==2.1.0
olefile==0.46
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.44
Pillow==5.4.1
proto-plus==1.13.0
protobuf==3.14.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyserial==3.4
pysmbc==1.0.15.6
python-apt==1.8.4.3
pytz==2020.5
pyxdg==0.25
PyYAML==5.4.1
rainbowhat==0.1.0
reportlab==3.5.13
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.7
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simplejson==3.16.0
six==1.15.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.4
ssh-import-id==5.7
thonny==3.3.3
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
unicornhathd==0.0.4
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
